fix: rename package name
