+++ docs/source/errors.rst
@@ -1,4 +1,4 @@
-.. currentmodule:: pandas_redshift
+.. currentmodule:: pandas_amazon_redshift
 .. _api.errors:
 
 **********
@@ -7,8 +7,8 @@ Exceptions
 
 .. note::
 
-   Only functions and classes which are members of the ``pandas_redshift``
-   and ``pandas_redshift.errors`` module are considered public.
+   Only functions and classes which are members of the ``pandas_amazon_redshift``
+   and ``pandas_amazon_redshift.errors`` module are considered public.
    Submodules and their members are considered private.
 
 .. autosummary::
