+++ pandas_amazon_redshift/core.py
