+++ pandas_amazon_redshift/errors.py
