+++ pandas_amazon_redshift/types.py
