+++ pandas_amazon_redshift/utils.py
