+++ setup.cfg
@@ -2,10 +2,10 @@
 [versioneer]
 VCS = git
 style = pep440
-versionfile_source = pandas_redshift/_version.py
-versionfile_build = pandas_redshift/_version.py
+versionfile_source = pandas_amazon_redshift/_version.py
+versionfile_build = pandas_amazon_redshift/_version.py
 tag_prefix =
-parentdir_prefix = pandas_redshift-
+parentdir_prefix = pandas_amazon_redshift-
 
 [flake8]
 exclude = docs
@@ -16,4 +16,4 @@ omit =
     # Omit the tests
     */tests/*
     # Omit generated versioneer
-    pandas_redshift/_version.py
+    pandas_amazon_redshift/_version.py
